requirements.txt
torch>=2.0.0
tokenizers>=0.13.0
datasets>=2.0.0
streamlit>=1.20.0
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
huggingface_hub>=0.20.0
